+++ src/lib/mina/v1/graphql.ts
@@ -1,7 +1,7 @@
 import { UInt32 } from '../../provable/int.js';
 import { TokenId, type ZkappCommand } from './account-update.js';
 import type { ActionStatesStringified } from './fetch.js';
-import { Types } from '../../../bindings/mina-transaction/types.js';
+import { Types } from '../../../bindings/mina-transaction/v1/types.js';
 
 export {
   type EpochData,