+++ src/lib/proof-system/zkprogram.ts
@@ -8,13 +8,8 @@ import {
 } from '../../snarky.js';
 import { Pickles, Gate } from '../../snarky.js';
 import { Field } from '../provable/wrapped.js';
-import {
-  FlexibleProvable,
-  InferProvable,
-  ProvablePureExtended,
-  Struct,
-} from '../provable/types/struct.js';
-import { InferProvableType, provable } from '../provable/types/provable-derivers.js';
+import { FlexibleProvable, InferProvable, ProvablePureExtended } from '../provable/types/struct.js';
+import { InferProvableType } from '../provable/types/provable-derivers.js';
 import { Provable } from '../provable/provable.js';
 import { assert, prettifyStacktracePromise } from '../util/errors.js';
 import { ConstraintSystemSummary, snarkContext } from '../provable/core/provable-context.js';
@@ -44,15 +39,17 @@ import { emptyWitness } from '../provable/types/util.js';
 import { From, InferValue } from '../../bindings/lib/provable-generic.js';
 import { DeclaredProof, ZkProgramContext } from './zkprogram-context.js';
 import { mapObject, mapToObject, zip } from '../util/arrays.js';
+import { VerificationKey } from './verification-key.js';
 
 // public API
-export { SelfProof, JsonProof, ZkProgram, verify, Empty, Undefined, Void, VerificationKey };
+export { SelfProof, JsonProof, ZkProgram, verify, Empty, Undefined, Void };
 
 // internal API
 export {
   CompiledTag,
   sortMethodArguments,
   MethodInterface,
+  MethodReturnType,
   picklesRuleFromFunction,
   compileProgram,
   analyzeMethod,
@@ -62,6 +59,7 @@ export {
   RegularProver,
   TupleToInstances,
   PrivateInput,
+  Proof,
 };
 
 type Undefined = undefined;
@@ -612,22 +610,6 @@ type ZkProgram<
  * ```
  */
 class SelfProof<PublicInput, PublicOutput> extends Proof<PublicInput, PublicOutput> {}
-
-class VerificationKey extends Struct({
-  ...provable({ data: String, hash: Field }),
-  toJSON({ data }: { data: string }) {
-    return data;
-  },
-}) {
-  static async dummy(): Promise<VerificationKey> {
-    await initializeBindings();
-    const [, data, hash] = Pickles.dummyVerificationKey();
-    return new VerificationKey({
-      data,
-      hash: Field(hash),
-    });
-  }
-}
 
 function sortMethodArguments(
   programName: string,