+++ src/mina-signer/tests/zkapp.unit-test.ts
@@ -1,5 +1,5 @@
-import { ZkappCommand } from '../../bindings/mina-transaction/gen/transaction-bigint.js';
-import * as TransactionJson from '../../bindings/mina-transaction/gen/transaction-json.js';
+import { ZkappCommand } from '../../bindings/mina-transaction/gen/v1/transaction-bigint.js';
+import * as TransactionJson from '../../bindings/mina-transaction/gen/v1/transaction-json.js';
 import Client from '../mina-signer.js';
 import { accountUpdateExample } from '../src/test-vectors/accountUpdate.js';
 import { expect } from 'expect';