+++ CHANGELOG.md
@@ -17,7 +17,14 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased](https://github.com/o1-labs/o1js/compare/fb625f...HEAD)
 
+### Added
+
 - Support of runtime tables https://github.com/o1-labs/o1js/pull/1858
+- _Experimental_ New bindings layer for new API types. https://github.com/o1-labs/o1js/pull/2032
+- _Experimental_ New API types for https://github.com/o1-labs/o1js/pull/2042
+  - `AccountUpdate`, `Account`, `Authorization`, `Permissions` etc.
+  - New transaction construction API `new ZkappCommand()`. https://github.com/o1-labs/o1js/pull/2042
+
 
 ## [2.3.0](https://github.com/o1-labs/o1js/compare/b857516...fb625f)