+++ CHANGELOG.md
@@ -23,6 +23,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - _Experimental_ New API types for https://github.com/o1-labs/o1js/pull/2042
   - `AccountUpdate`, `Account`, `Authorization`, `Permissions` etc.
   - New transaction construction API `new ZkappCommand()`. https://github.com/o1-labs/o1js/pull/2042
+- Bump up Rust version to 1.79.0. Bindings now depends on nightly-2024-06-13.
+  https://github.com/o1-labs/o1js/pull/2063
 
 ## [2.3.0](https://github.com/o1-labs/o1js/compare/b857516...fb625f)